Avoid gcc compiler warnings related to sprintf format

+++ operator.c
@@ -700,9 +700,9 @@ static void opCmdEnterKeys(bool help, char *cmdParams)
      */
     clock = time(NULL);
     tmp = localtime(&clock);
-    sprintf(timestamp, "%02.2d%02.2d%02.2d%02.2d%02.2d%02.2d",
-        tmp->tm_year - 100, tmp->tm_mon + 1, tmp->tm_mday,
-        tmp->tm_hour, tmp->tm_min, tmp->tm_sec);
+    sprintf(timestamp, "%02d%02d%02d%02d%02d%02d",
+        (u8)tmp->tm_year - 100, (u8)tmp->tm_mon + 1, (u8)tmp->tm_mday,
+        (u8)tmp->tm_hour, (u8)tmp->tm_min, (u8)tmp->tm_sec);
     cp = cmdParams;
     bp = keybuf;
     limit = bp + sizeof(keybuf) - 2;
